Drag drop perfectly- Stable version

+++ app/Controller/UsersController.php
@@ -5,11 +5,27 @@ Class UsersController extends AppController
     var $components = array('Email');
 	var $name='Users';  
 	public $uses = array('User');
-	public function beforeFilter()
-	{		
-	}
+	public function beforeFilter() {
+        parent::beforeFilter();
+        // load all admin template
+		$cuser = $this->Session->read('user');
+		$this->loadModel('Sheet');
+		$this->Sheet->unbindModel(array('hasMany' => array('SheetPage'), 'hasOne' => array('TeacherPage')));
+		$templates = array();
+		if ($cuser['User']['user_role'] == 'teacher') {
+			$templates = $this->Sheet->find('all', array(
+				'conditions' => array('Sheet.type' => 'admin', 'Sheet.template' => 'teacher')
+			));
+		} else {
+			$templates = $this->Sheet->find('all', array(
+				'conditions' => array('Sheet.type' => 'admin', 'Sheet.template' => $cuser['User']['user_role'])
+			));
+		}
+		$this->set(compact('templates'));
+    }
 	public function index()
 	{
+		$this->redirect('/users/dashboard');
 	}
 	public function forgetpassword()
 	{
@@ -119,21 +135,48 @@ public function changepassword($email='')
 		}
 		
 	}
-	public function dashboard()
+	public function dashboard($editWorkSheet = null)
 	{	
-		$this->layout='math';
 		$userData = $this->Session->read('user');
                 if(sizeof($userData)>0){
                 	// biplob start
-                	$this->Session->delete('Sheet');
-					$this->Session->delete('SheetPage');
-                	// check if form post  
-                	if ($this->request->is(array('post','put'))) {
-                		$this->Session->write('Sheet.name', $this->request->data['Sheet']['name']);
-	                }
-	       			// load all saved sheet
+                	if ($userData['User']['user_role'] == 'teacher') {
+                		$this->layout='teacher';
+                	} else {
+                		$this->layout='math';
+                	}
+                	if ($editWorkSheet) {
+						$review_data = $this->Session->read('SheetPage');
+						// if session expired
+						if (empty($review_data)) {
+							$this->Session->delete('Sheet');
+							$this->Session->delete('SheetPage');
+							return $this->redirect(array('controller' => 'users', 'action' => 'dashboard'));
+						}
+						$this->set('questions', json_decode($review_data[0], true));
+						$this->set('page_count', count($review_data));
+					} else {
+						$this->Session->delete('Sheet');
+						$this->Session->delete('SheetPage');
+					}
+					// load all saved sheet
 					$this->loadModel('Sheet');
 					$this->Sheet->unbindModel(array('hasMany' => array('SheetPage'), 'hasOne' => array('TeacherPage')));
+                	// check if form post  
+                	if ($this->request->is(array('post','put'))) {
+                		// sheet name validation
+                		$sheetNameExist = $this->Sheet->find('first', array(
+                			'conditions' => array(
+                				'Sheet.name' => $this->request->data['Sheet']['name'],
+                				'Sheet.user_id' => $userData['User']['id']
+                			)
+                		));
+                		if (!empty($sheetNameExist)) {
+                			$this->Session->setFlash(__('You have created a worksheet using the same name, please choose a unique name.'), 'error_form', array(), 'error');
+                		} else {
+                			$this->Session->write('Sheet.name', $this->request->data['Sheet']['name']);
+                		}
+	                }	       			
 					$sheets = $this->Sheet->find('all', array(
 						'conditions' => array('Sheet.user_id' => $userData['User']['id']),
 						'fields' => array('Sheet.id', 'Sheet.name')
@@ -180,7 +223,7 @@ else{echo "<li><a href='#'>No Topics Found</a></li>";}
  }
 echo "</ul>";
 }
-else{echo "<li><a href='#'>No Topics Found</a></li>";}
+else{echo "<li><a href='#'>No Questions Found</a></li>";}
 		exit;
 		  
 	}
@@ -360,22 +403,90 @@ else{echo "<li><a href='#'>No Topics Found</a></li>";}
 		$c_user = $this->Session->read('user');
 		if ($c_user['User']['user_role'] == 'teacher') {
 			$this->Session->write('SheetPage.0', $sheetData);
-		} 
+		} else {
+			$this->Session->write('SheetPage.' . $_REQUEST['page_index'], $sheetData);
+		}
+	}
+
+	/*
+	* Math specialist page add
+	*/
+	public function ajax_page_add() {
+		$this->layout = 'ajax';
+		$review_data = $this->Session->read('SheetPage');
+		if (isset($this->request->data['review'])) {
+			$this->set('review', true);
+		}
+		if (isset($review_data[$this->request->data['page_no']])) {
+			$this->set('questions', json_decode($review_data[$this->request->data['page_no']], true));		
+		} else {
+			$this->set('questions', array());
+		}
+		
+		$this->render('/Elements/users/edit_sheet');
+	}
+
+	/*
+	* Math specialist page delete
+	*/
+	public function ajax_page_delete() {
+		$this->layout = 'ajax';
+		$currentPage = $this->request->data['page_no'];
+		$maxPage = $this->request->data['max_page'];
+
+		$this->Session->delete('SheetPage.' . ($currentPage - 1));
+
+		if (($currentPage >= 1) && ($currentPage < $maxPage)) {
+          // any page, not one page and not the last page
+			$review_data = $this->Session->read('SheetPage');
+			$i = 0;
+            foreach ($review_data as $key => $value) {
+            	$this->Session->write('SheetPage.' . $i, $value);
+            	$i++;
+            }
+            $this->set('questions', json_decode($review_data[$currentPage], true));
+        } elseif (($currentPage > 1) && ($currentPage == $maxPage)) {
+            // if last page delete
+        	$review_data = $this->Session->read('SheetPage');
+            $this->set('questions', json_decode($review_data[$currentPage-2], true));
+        } else {
+        	// if one page and delete
+        	$this->set('questions', array());
+        }
+		$this->render('/Elements/users/edit_sheet');
 	}
 
 	/*
 	* Teacher page review
 	*/
-	public function review() {
+	public function review($template_id = null) {
 		$this->layout = 'review';
-		$review_data = $this->Session->read('SheetPage');
-		if (empty($review_data)) {
-			return $this->redirect(array('controller' => 'users', 'action' => 'dashboard'));
-		}
-		$this->set('questions', json_decode($review_data[0], true));
-		$this->set('page_count', count($review_data));
-
 		$userData = $this->Session->read('user');
+		if (empty($template_id)) {
+			$review_data = $this->Session->read('SheetPage');
+			if (empty($review_data)) {
+				return $this->redirect(array('controller' => 'users', 'action' => 'dashboard'));
+			}
+			$this->set('questions', json_decode($review_data[0], true));
+			$this->set('page_count', count($review_data));
+		} else {
+			// delete session
+			$this->Session->delete('Sheet');
+			$this->Session->delete('SheetPage');
+			$sheet = $this->Sheet->findById($template_id);
+			$this->Session->write('Sheet.name', $sheet['Sheet']['name']);
+			if (!empty($sheet['SheetPage'])) {
+				foreach ($sheet['SheetPage'] as $key => $value) {
+					$this->Session->write('SheetPage.' . $key, $value['params']);
+				}
+				$this->set('questions', json_decode($sheet['SheetPage'][0]['params'], true));
+				if ($sheet['Sheet']['template'] != 'teacher') {
+					$this->set('page_count', count($sheet['SheetPage']));
+				}
+			}
+			$this->set('admin_template', true);
+		}
+
 		if(sizeof($userData)>0){
 			$userDetail = $this->User->find('first',array('conditions'=>array('User.id'=>$userData['User']['id'])));        
 			$this->set('userDetail',$userDetail);
@@ -390,21 +501,32 @@ else{echo "<li><a href='#'>No Topics Found</a></li>";}
 	*/
 	public function ajax_savepage(){
 		$this->autoRender = false;
-		if (!$this->Session->check('Sheet.id')){
-			$c_user = $this->Session->read('user');
-			$sheetPageData = $this->Session->read('SheetPage');
-			// if ($this->Session->check('Sheet.id')){
-			// 	$sheetData['Sheet']['id'] = $this->Session->read('Sheet.id');
-			// 	$sheetData['TeacherPage']['sheet_id'] = $this->Session->read('Sheet.id');
-			// }
-			$sheetData['Sheet']['name'] = $this->Session->read('Sheet.name');
-			$sheetData['Sheet']['user_id'] = $c_user['User']['id'];
-			$sheetData['Sheet']['type'] = $c_user['User']['user_role'];
-			$sheetData['TeacherPage']['params'] = $sheetPageData[$this->request->data['save_index']];
-			$this->loadModel('Sheet');
-			$this->Sheet->saveAll($sheetData);
-			$this->Session->write('Sheet.id', $this->Sheet->id);
+		$this->loadModel('Sheet');
+		$c_user = $this->Session->read('user');
+		$sheetPageData = $this->Session->read('SheetPage');
+		if ($this->Session->check('Sheet.id')){
+			$sheetData['Sheet']['id'] = $this->Session->read('Sheet.id');
+			if ($c_user['User']['user_role'] == 'teacher') {
+				$sheetData['TeacherPage']['sheet_id'] = $this->Session->read('Sheet.id');
+				$c_sheetPage = $this->Sheet->SheetPage->findBySheetId($sheetData['Sheet']['id']);
+				$sheetData['TeacherPage']['id'] = $c_sheetPage['SheetPage']['id'];
+			} else {
+				$this->Sheet->SheetPage->deleteAll(array('SheetPage.sheet_id' => $sheetData['Sheet']['id']));
+			}
 		}
+		if ($c_user['User']['user_role'] == 'teacher') {
+			$sheetData['TeacherPage']['params'] = $sheetPageData[$this->request->data['save_index']];
+		} else {
+			foreach ($sheetPageData as $key => $value) {
+				$sheetData['SheetPage'][$key]['params'] = $value;
+			}
+		}
+		$sheetData['Sheet']['name'] = $this->Session->read('Sheet.name');
+		$sheetData['Sheet']['user_id'] = $c_user['User']['id'];
+		$sheetData['Sheet']['type'] = $c_user['User']['user_role'];
+		$this->Sheet->saveAll($sheetData);
+		$this->Session->write('Sheet.id', $this->Sheet->id);
+		
 	}
 
 	/*
@@ -416,8 +538,50 @@ else{echo "<li><a href='#'>No Topics Found</a></li>";}
 		if (empty($review_data)) {
 			return $this->redirect(array('controller' => 'users', 'action' => 'dashboard'));
 		}
-		$this->set('questions', json_decode($review_data[$this->request->data['save_index']], true));
-		$this->set('page_count', count($review_data));
+		$c_user = $this->Session->read('user');
+
+		if (!empty($this->request->data['spanish']) && empty($this->request->data['answerkey'])) {
+			if ($c_user['User']['user_role'] == 'teacher') {
+				$this->set('questions', json_decode($review_data[$this->request->data['save_index']], true));
+			} else {
+				foreach ($review_data as $key => $value) {
+					$all_pages[$key] = json_decode($value, true);
+				}
+				$this->set(compact('all_pages'));
+			}
+
+		} elseif (empty($this->request->data['spanish']) && empty($this->request->data['answerkey'])) {
+			if ($c_user['User']['user_role'] == 'teacher') {
+				$this->set('questions', json_decode($review_data[$this->request->data['save_index']], true));
+			} else {
+				foreach ($review_data as $key => $value) {
+					$all_pages[$key] = json_decode($value, true);
+				}
+				$this->set(compact('all_pages'));
+			}
+
+		} elseif (!empty($this->request->data['spanish']) && !empty($this->request->data['answerkey'])) {
+			if ($c_user['User']['user_role'] == 'teacher') {
+				$questions = json_decode($review_data[$this->request->data['save_index']], true);
+				$questions = array_merge($questions, $questions);
+				$this->set(compact('questions'));
+			} else {
+				foreach ($review_data as $key => $value) {
+					$all_pages[$key] = json_decode($value, true);
+				}
+				$all_pages = array_merge($all_pages, $all_pages);
+				$this->set(compact('all_pages'));
+			}
+		} else {
+			if ($c_user['User']['user_role'] == 'teacher') {
+				$this->set('questions', json_decode($review_data[$this->request->data['save_index']], true));
+			} else {
+				foreach ($review_data as $key => $value) {
+					$all_pages[$key] = json_decode($value, true);
+				}
+				$this->set(compact('all_pages'));
+			}
+		}
 	}
 
 	/*
@@ -444,14 +608,21 @@ else{echo "<li><a href='#'>No Topics Found</a></li>";}
 		} else {
 			$this->Sheet->unbindModel(array('hasOne' => array('TeacherPage')));
 		}
-		$sheet = $this->Sheet->findByIdAndType($this->request->data['id'], $c_user['User']['user_role']);
+		$sheet = $this->Sheet->findById($this->request->data['id']);
 		$this->Session->delete('Sheet');
 		$this->Session->delete('SheetPage');
 		$this->Session->write('Sheet.id', $sheet['Sheet']['id']);
 		$this->Session->write('Sheet.name', $sheet['Sheet']['name']);
 		// for teacher
 		if (!empty($sheet['TeacherPage']['params'])) {
+			$this->Session->write('SheetPage.0', $sheet['TeacherPage']['params']);
 			$questions = json_decode($sheet['TeacherPage']['params'], true);
+		} elseif (!empty($sheet['SheetPage'])) {
+			$questions = json_decode($sheet['SheetPage'][0]['params'], true);
+			$this->set('page_count', count($sheet['SheetPage']));
+			foreach ($sheet['SheetPage'] as $key => $value) {
+				$this->Session->write('SheetPage.' . $key, $value['params']);
+			}
 		} else {
 			$questions = array();
 		}
@@ -484,6 +655,34 @@ else{echo "<li><a href='#'>No Topics Found</a></li>";}
 			$used_questions = array();
 		}
 		$this->set(compact('questions', 'used_questions'));
+	}
+
+	/*
+	* ajax load grade drop down
+	*/
+	public function ajax_grades_dropdown() {
+		$this->layout = 'ajax';
+		$this->loadModel('GradeLevel');
+		$query="select * from grade_levels";
+        $gradeLevel= $this->User->query($query);
+     	$gradeLevels = $this->GradeLevel->find('all');
+        $this->set(compact('gradeLevels'));
+	}
+
+	/*
+	* ajax search by image index no
+	*/
+	public function ajax_search_image() {
+		$this->layout = 'ajax';
+		$this->loadModel('Question');
+		// get all questions
+		$questions = $this->Question->find('all', array(
+			'conditions' => array(
+				"Question.index_no LIKE" => '%' . $this->request->data['search_index'] . '%',
+			),
+			'fields' => array('Question.id', 'Question.file_name', 'Question.size')
+		));
+        $this->set(compact('questions'));
 	}
